#942 Improve scope break down process
OpenGnsys has a hierarchical scope system to organize clients:

				 +------+
				 |Center|
				 +------+
				 |      |
				 |      |
				 v      v
	    +--------------------+      +----------+
	+-->+Groups of classrooms+----->+Classrooms|
	|   +-+------------------+      +----------+
	|     |                         |          |
	+-----+                         |          |
					v          v
		      +-----------------+          +-------+
		  +-->+Groups of clients+--------->+Clients|
		  |   +-+---------------+          +-------+
		  |     |
		  +-----+

OpenGnsys can send commands and tasks to several clients simultaneously
selecting a scope. For example: send poweroff to center Faculty of
Pharmacy. We had a waterfall process to loop through all members of the
scope and obtain the clients.

This commit simplifies the process with queries that save us to loop
through the members.

Co-authored-by: Irina Gómez <irinagomez@us.es>