#1004 Fix client setup disk size showing as 0
og_cmd_get_client_setup is querying the database for computers disk
setup information. This includes the size of a disk or a partition,
which are stored as BIGINT in the database.

Using dbi_result_get_int will result in an error as libdbi expects to
store this kind of data type in a long long instead of an int.

Fix disk size variable declaration and use the correct dbi_result_get
function for this particular data column.