#990 wol: migrate mac and netmask query to ogServer
ogServer WoL rest function is expecting ip, mac and netmask from the
request's payload. This makes the client responsible for providing such
data.

ogServer should be the owner of the database, this patch adjusts the
parameter expected in a request payload for /wol action so clients only
need to provide the ips and wol type (broadcast, unicast).

Database is expected to contain valid data for the netmask and mac of
target computers.