#1019 Fix queued Wake on LAN
UMA and UPV report that Wake on LAN command (in queue mode) does not
work.

We improved WoL command, now ogServer calculates the broadcast address
of the network to which the client belongs. To calculate this address
ogServer needs the IP and the netmask of the client. We updated ogServer
to retrieve the netmask from the database in non-queue mode, but we forgot
to add this in queue mode.

This patch adds netmask retrieving to queued WoL.