#997 Use stale check flag in schedule/update
Commit e68fefe introduced 'check_stale' flag to better distinguish
real scheduled actions that do not execute if they are stale from
immediate actions that we want them to be logged in the action queue
(by creating a decoy schedule for the exact moment they are processed,
meaning that we ignore if the are stale).

Add this feature into schedule update too, in order to avoid executing
stale commands that were not meant to, ie. real scheduled commands.

Follows e68fefe ("Set stale check flag when processing schedule/create")