schedule: fix daylight saving problem with mktime()
mktime modifies the struct tm it receives and takes into account whether DST is
active or not (tm_isdst). tm_isdst == 0 adjusts the time, which causes the time
mismatch error.

All fields are being initialized to 0 and therefore it is assumed that the time
that has been passed is not in daylight saving time.

When the value is negative in tm.tm_isdst it delegates to mktime to guess if it
is in daylight saving time or not, this works 99% of the time.

Best way would be that ogserver knows what is its timezone and when daylight
saving applies, so tm_isdst is set to 0 or 1 accordingly.

Meanwhile, "tm_isdst = -1" provides the hotfix.