#1054 Fix og_legacy_partition code buffer size
The "code" member of the og_legacy_partition is used to hold the string of
the partition type *name* for legacy parameter strings. Example:

par=2*cpt=LINUX-SWAP*sfi=LINUX-SWAP*tam=10000000*ope=0

Problem: Buffer size is smaller than possible values for this member.
Fix: Increase buffer size up to the max length defined in the DB
for the corresponding column.